+++ src/factory/createStateContext.ts
@@ -5,7 +5,13 @@ const createStateContext = <T>(defaultInitialValue: T) => {
     createContext<[T, React.Dispatch<React.SetStateAction<T>>] | undefined>(undefined);
   const providerFactory = (props, children) => createElement(context.Provider, props, children);
 
-  const StateProvider: React.FC<{ initialValue?: T }> = ({ children, initialValue }) => {
+  const StateProvider = ({
+    children,
+    initialValue,
+  }: {
+    children?: React.ReactNode;
+    initialValue?: T;
+  }) => {
     const state = useState<T>(initialValue !== undefined ? initialValue : defaultInitialValue);
     return providerFactory({ value: state }, children);
   };