+++ stories/useHarmonicIntervalFn.story.tsx
@@ -3,7 +3,12 @@ import * as React from 'react';
 import { useHarmonicIntervalFn, useInterval, useTimeoutFn } from '../src';
 import ShowDocs from './util/ShowDocs';
 
-const Clock: React.FC<{ useInt: typeof useHarmonicIntervalFn }> = ({ useInt }) => {
+const Clock = ({
+  useInt,
+}: {
+  children?: React.ReactNode;
+  useInt: typeof useHarmonicIntervalFn;
+}) => {
   const [count, setCount] = React.useState(0);
   useInt(() => {
     setCount((cnt) => cnt + 1);
@@ -25,7 +30,7 @@ const Clock: React.FC<{ useInt: typeof useHarmonicIntervalFn }> = ({ useInt }) =
   return <div style={style}>{m + ':' + s}</div>;
 };
 
-const Demo: React.FC<{ useInt: typeof useHarmonicIntervalFn }> = ({ useInt }) => {
+const Demo = ({ useInt }: { children?: React.ReactNode; useInt: typeof useHarmonicIntervalFn }) => {
   const [showSecondClock, setShowSecondClock] = React.useState(false);
   useTimeoutFn(() => {
     setShowSecondClock(true);